Extract the Codex config codec into cli/integrations/
Move the Codex ~/.codex/config.toml codec out of the setup command module
and into its own cli/integrations module, alongside the other surface
handlers. _default_codex_config_path, the _CodexNauroEntry typed view,
_parse_codex_nauro_entry, and _configure_codex form a self-contained
load/parse/mutate/write unit with no dependency on the setup command
surface, so setup.py should not host them. This is a verbatim move with no
behavior change: setup.py imports _configure_codex and
_default_codex_config_path, the two symbols it still calls directly, and
drops the tomlkit imports the codec was the only user of.
